🍻 Add new circles to Memory Game
